requirements.txt
# Core Flask and Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# AI and Machine Learning
google-generativeai==0.3.2

sentence-transformers>=2.7.0
transformers>=4.38.0
torch>=2.2.0
numpy==1.24.4
scikit-learn==1.3.2
pandas==2.1.4

# Vector Database and Embeddings
chromadb==0.4.18
langchain==0.0.350
langchain-community==0.0.4
langchain-google-vertexai
faiss-cpu==1.7.4

# PDF Processing and OCR
PyPDF2==3.0.1
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3

# Text Processing and NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
langdetect==1.0.9
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration and Data Processing
PyYAML==6.0.1
python-dotenv==1.0.0
pydantic==2.5.2
marshmallow==3.20.1
jsonschema==4.20.0

# Database and Storage
SQLAlchemy==2.0.22
redis==5.0.1
pymongo==4.6.0
pymemcache==4.0.0

# HTTP and API
requests==2.31.0
urllib3==2.1.0
httpx==0.13.3

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and Monitoring
loguru==0.7.2
sentry-sdk[flask]==1.38.0

# Security
cryptography==41.0.4

bcrypt==4.1.2
PyJWT==2.8.0

# Performance and Caching
gunicorn==21.2.0
gevent==23.9.1

# Data Validation and Serialization
cerberus==1.3.5
schema==0.7.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
rich==13.7.0
colorama==0.4.6

# Environment and Deployment
python-decouple==3.8
whitenoise==6.6.0

# Multi-language Support
googletrans==4.0.0rc1

pycountry==23.12.11

# File Handling
pathlib2==2.3.7
watchdog==3.0.0

# Image Processing (for OCR)
imageio==2.33.0
matplotlib==3.8.2

# Progress and UI
alive-progress==3.1.5

# Production WSGI Server
waitress==2.1.2

# Additional utilities for production
psutil==5.9.6
schedule==1.2.0
